% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Plots the accidents on a map for specified state and year}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{Input is a numerical value which references a state}

\item{year}{Input should be numerical value denoting year.}
}
\description{
This function takes 'state number' and 'year' as input and verifies if its a valid state
number and then check if the number of accidents is NONE or not. If not then reads the
latitude and longitude and plots it on graphical maps
}
\note{
It throws an error if state number is not valid and prints"invalid state number" state.num"
 It returns a message "no accidents found" if number of accidents for specified state and year is 0.
}
\examples{
fars_map_state(1,2015)
fars_map_state('1','2015')
fars_map_state("1","2015")

}
