% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Import Dataset - US NHTSA Fars Data (Fatality Analysis Reporting System)}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string specifying just the filename if it resides in current
working directory, else,  specifying the absolute path to the file.}
}
\value{
This function returns DataFrame tbl class for the dataset.This function stops
        executing if the file does not exits in the specified directory and throws error
        message "file <filename> does not exists".
}
\description{
This function check the data is available in the specified directory and
then imports it to global environment by read it in .CSV format.
}
\examples{
fars_read(make_filename(2013))

}
