% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Selects the data based on 'month' an saves as separate file extension - .csv.bz2 file}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{Takes number as input which will be used in formatted string literal.
Would throw an error if alphabets is provided as input argument.}
}
\value{
A compressed file containing dataset for the the specified month.
}
\description{
This function uses base R function 'lapply' and also uses function - 'mutate', 'select'
from "dplyr" package. Takes years as arguments, create a formatted string object used for
naming the saved file and finally, selects the 'MONTH' and 'year' column and assigns it to
"dat" object. Returns error if the "year" is not found.
}
\note{
If the specified year is not there in the dataset, It throws a warning: "invalid year: <year>" and
 returns NULL. It would return NULL in case of alphbet being passed as an argument.
}
\examples{
fars_read_years(2015)

}
