% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Prints "accident_<year>.csv.bz2"}
\usage{
make_filename(year)
}
\arguments{
\item{year}{Takes number as input which will be used in formatted string literal. Would throw an error
if alphabets is provided as input argument}
}
\value{
returns formatted string as output.
}
\description{
This function takes the year as input an prints the formatted string "accident_year.csv.bz2"
}
\examples{
make_filename(2013)

}
